[Design/#214] 파일 분리(QuestStyle) ,뷰 이름 변경(QuestStatHeaderView >> StatHeaderView)
