+++ ILSANG/Sources/Views/Quest/QuestView.swift
@@ -73,7 +73,7 @@ extension QuestView {
     
     // 서브헤더 - 5가지 스탯
     private var subHeaderView: some View {
-        QuestStatHeaderView(
+        StatHeaderView(
             selectedXpStat: $vm.selectedXpStat,
             horizontalPadding: 0,
             height: 44,
